Update README; work toward three-examples
